requirements.txt
alembic==1.0.0
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mock==2.0.0
more-itertools==4.2.0
pbr==4.2.0
pluggy==0.6.0
py==1.5.4
PyJWT==1.6.4
pytest==3.6.3
python-dateutil==2.7.3
python-editor==1.0.3
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.10
urllib3==1.23
Werkzeug==0.14.1
